+++ src/App.js
@@ -1,8 +1,7 @@
 const MissionUtils = require("@woowacourse/mission-utils");
-const validateInput = require("./ValidateInput");
-const { getStrikeAndBall, getStrikeAndBallText } = require("./StrikeAndBall");
 const getThreeRandomNumbers = require("./ThreeRandomNumbers");
-const { NUMBER_LENGTH, END_INPUT, RESTART_INPUT } = require("./constants/ConstantValues");
+const playGame = require("./PlayGame");
+const { END_INPUT, RESTART_INPUT } = require("./constants/ConstantValues");
 const {
   START_MESSAGE,
   INPUT_NUMBER_MESSAGE,
@@ -18,7 +17,7 @@ class App {
   }
 
   play() {
-    MissionUtils.Console.readLine(INPUT_NUMBER_MESSAGE, this.playGame.bind(this));
+    MissionUtils.Console.readLine(INPUT_NUMBER_MESSAGE, playGame.bind(this));
   }
 
   end() {
@@ -29,23 +28,6 @@ class App {
   throwError() {
     MissionUtils.Console.close();
     throw new Error(WRONG_INPUT_ERROR_MESSAGE);
-  }
-
-  playGame(input) {
-    if (!validateInput(input)) {
-      this.throwError();
-    }
-
-    const [strikeCount, ballCount] = getStrikeAndBall(this.threeRandomNumbers, input);
-    const resultOutput = getStrikeAndBallText(strikeCount, ballCount);
-
-    MissionUtils.Console.print(resultOutput);
-
-    if (strikeCount === NUMBER_LENGTH) {
-      this.end();
-    }
-
-    this.play();
   }
 
   endGame(input) {
